Make a lack of POC information an error that can be overridden
By default an INI file with no POC information or an INI with at least
one POC with an empty string for an email address will both be
rejected by cyhy-simple.  However, the --force command line option can
be used to override the error and force creation of a JSON file.