Update geoipupdate.sh script to account for MaxMind's change in free db access.
This updates the script primarily to accomodate for MaxMind's changed policy in
how to get a copy of GeoLite2, the free offering of their GeoIP database. They
now require a license key for any version and the previous URL used for the
free version is no longer valid. I also performed some housekeeping on the
script to flesh it out more.