Update README and add helper script for CyHy Team Docker image creation.
The README has been updated with instructions that take into account the changes
made to the `var/geoipupdate.sh` script. A new helper script to create an
appropriate image and save it locally has also been included. The .gitignore and
.dockerignore files have been updated to reflect the addition of the helper
script, and the .git directory has been added to .dockerignore to clean up
generated images.