Fix language in generate_cyhy_docker_image.sh and update README examples.
Correct the language in generate_cyhy_docker_image.sh to correctly label the
two parts of the image descripton as name and tag (instead of tag and version).
Updated the generate_cyhy_docker_image.sh examples in the README to show how you
can optionally specify the image name and tag, or just the tag.