Update Docker image script to pull license from AWS SSM
Previously this relied on pulling the MaxMind GeoIP2 license key from SSM
and then adding it to a file. This file was then pulled in to populate the
appropriate Docker build-arg. This currently uses awk, which may be fragile
if AWS changes the response structure. Using jq and switching to
`--output json` to be more resilient to small API changes.