Correct CVSS validation logic in the cyhy-nvdsync script
The current check to validate CVSS data will not provide accurate
results if the the list it is checking has only one element. This
results in any modern CVEs in the feed, which generally only have a
`baseMetricV3` key, failing the validation step. As a result only a
single CVE from the 2023 feed (out of 3,228 total CVE entries) is
imported with the current logic.