Add the re-ID script
This adds the id_update.py script we use to re-ID stakeholders to this
repository. Until this is either absorbed into an existing CyHy command
or turned into its own command we will at least have the script as part
of the regular codebase. This is preferable to storing it in the
private oneoffs repository even though it contains nothing sensitive.