Check for non-ASCII characters in input to cyhy-simple
Currently if non-ASCII characters are encounted in input to the
cyhy-simple script they are converted to a Unicode escape sequence. The
escape sequence itself is comprised of ASCII characters which allows it
to be processed safely by the cyhy-import script, however it is
interpolated to the represented Unicode character by MongoDB. This
results in breaking Unicode characters being present in MongoDB
documents.