Always import data in the GNIS data importer script
This removes the `--force` option and any existing data checks. The
script will instead always load GNIS data on every run. This ensures
that no data is lost as the existing data check only looked at the
first and last records in the incoming file. If any records were
modified or added between the first and last records then they would
not be updated/imported. Since this script is typically only run on
instance redeployment in our production environment there is not much
gain for checking for existing records.

Also of note is that with the merge of #83 the check for existing
records will now throw an error. This is because the csv.DictReader
object begins reading the file (to pull the CSV header) before the
checking logic is run. This results in the following error:
IOError: telling position disabled by next() call