Provide an override for restricted IPs in cyhy-import
There are cases where restricted IPs may need to be added to the
database. Therefore, we provide a flag to override the default behavior
that disallows restricted IPs. Although the odds of needing this
functionality in the cyhy-import tool are slim, we add it for feature
consistency with the changes made to the cyhy-ip tool.