Define a default owner
This owner's request document will be created when the system initializes.  Initial use is to own all "ownerless" HostDocs, which are created when a CyHy entity has a hostname that resolves to an IP address not already owned by a CyHy entity.  This is how we account for cases where an entity owns a hostname, but not necessarily the IP addresses that it resolves to.