Update tracked issue in a comment
In the new workflow I copied over content from cisagov/ncats-webui that
was similar to the needs in this project. That included a comment block
that references an issue in that repository. I have updated the comment
to refer to a newly-created issue in this repository for better
tracking/visibility.