If hostname field is not present (i.e. tickets created before hostnames were supported), set hostname to None (as opposed to the "" empty string)
We don't want to set the hostname to "" in case this ticket is saved back to the DB (i.e. when modifying the false_positive flag) because the commander could potentially create a new ticket whose only difference would be the hostname set to None and we want to avoid that situation.